refactor(certificates): share nginx reload logic
ACME and get-cert maintained identical nginx process discovery and reload logic. A shared helper keeps certificate reload behavior consistent while removing duplicate code and tests.

- Route certificate installation, renewal, and watched-file reloads through cmdsutil.ReloadNginx.
- Preserve procfs master matching, SIGHUP delivery, error wrapping, and each caller's logger.
- Consolidate discovery and signal tests while retaining both commands' integration coverage.
