+++ internal/cmds/getcert/run.go
@@ -81,8 +81,7 @@ type config struct {
 // the control plane cannot redirect).
 var inventoryEndpoint = workloadclaims.InventoryEndpoint
 
-// procRoot is the procfs mount findNginxMasterPID scans. It is a package
-// variable only so tests can substitute a fake /proc tree.
+// procRoot is the procfs mount used to find nginx; tests substitute a fake tree.
 var procRoot = "/proc"
 
 var (
@@ -360,7 +359,7 @@ func renewLoop(ctx context.Context, cfg config, client attestclient.Client, leaf
 				unnamedRuns++
 			}
 			if cfg.ReloadNginx {
-				if err := reloadNginx(); err != nil {
+				if err := cmdsutil.ReloadNginx(procRoot, slog.Default()); err != nil {
 					slog.Warn("certificate renewed but nginx reload failed", "error", err)
 				}
 			}
@@ -376,7 +375,7 @@ func renewLoop(ctx context.Context, cfg config, client attestclient.Client, leaf
 				continue
 			}
 			slog.Info("watched file changed, reloading nginx")
-			if err := reloadNginx(); err != nil {
+			if err := cmdsutil.ReloadNginx(procRoot, slog.Default()); err != nil {
 				slog.Warn("watched file changed but nginx reload failed", "error", err)
 			}
 		}
@@ -614,58 +613,6 @@ func fetchSandboxToken(ctx context.Context, cfg config, pub crypto.PublicKey, no
 		return nil, fmt.Errorf("marshal sandbox token: %w", err)
 	}
 	return raw, nil
-}
-
-// reloadNginx sends SIGHUP to the nginx master process to reload certs.
-// Requires shareProcessNamespace: true in the pod spec. Walks /proc directly
-// instead of shelling out to pgrep so this works in distroless images.
-func reloadNginx() error {
-	pid, err := findNginxMasterPID()
-	if err != nil {
-		return err
-	}
-	proc, err := os.FindProcess(pid)
-	if err != nil {
-		return fmt.Errorf("find process %d: %w", pid, err)
-	}
-	if err := proc.Signal(syscall.SIGHUP); err != nil {
-		return fmt.Errorf("SIGHUP nginx (pid %d): %w", pid, err)
-	}
-	slog.Info("sent SIGHUP to nginx", "pid", pid)
-	return nil
-}
-
-// findNginxMasterPID scans /proc for the nginx master process.
-// Match: /proc/<pid>/comm == "nginx" AND cmdline contains "master".
-func findNginxMasterPID() (int, error) {
-	entries, err := os.ReadDir(procRoot)
-	if err != nil {
-		return 0, fmt.Errorf("read %s: %w", procRoot, err)
-	}
-	for _, e := range entries {
-		if !e.IsDir() {
-			continue
-		}
-		pid, err := strconv.Atoi(e.Name())
-		if err != nil {
-			continue
-		}
-		comm, err := os.ReadFile(procRoot + "/" + e.Name() + "/comm")
-		if err != nil || strings.TrimSpace(string(comm)) != "nginx" {
-			continue
-		}
-		cmdline, err := os.ReadFile(procRoot + "/" + e.Name() + "/cmdline")
-		if err != nil {
-			continue
-		}
-		// /proc/<pid>/cmdline is NUL-separated; nginx master argv[0] is
-		// "nginx: master process ...".
-		if !strings.Contains(string(cmdline), "master") {
-			continue
-		}
-		return pid, nil
-	}
-	return 0, fmt.Errorf("no nginx master process found")
 }
 
 // validateConfig checks that all required configuration is valid.
